Server/requirements.txt
Deprecated==1.2.13
grpcio==1.51.1
grpcio-tools==1.51.1
joblib==1.2.0
llvmlite==0.39.1
numba==0.56.4
numpy==1.23.5
pandas==1.5.3
protobuf==4.22.0
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sktime==0.16.1
threadpoolctl==3.1.0
wrapt==1.14.1
